Bring footer_t up to match new templated footer

+++ footer_t.php
@@ -17,7 +17,7 @@
 //
 // File: footer_t.php
 
-$online = 0;
+$online = (integer) 0;
 
 if (!$db->inactive)
 {
@@ -139,7 +139,6 @@ else
     $sf_logo_link = "?lang=" . $_GET['lang'];
 }
 
-//$elapsed = number_format ($elapsed, 2);
 $mem_peak_usage = floor (memory_get_peak_usage() / 1024);
 
 $public_pages = array ( 'ranking.php', 'new.php', 'faq.php', 'settings.php', 'news.php', 'index.php');
